Huge cleanup and simplification
The whole architecture just got way easier, it can be boiled down to "Client app asks KPeopleManager to fetch the data, KPM tells PersonCache to perform the query and store the persons and create a tree model out of them, KPM then returns this model, which can be filtered by the type of contact data it contains".

As there are more queries or more persons added to the cache (and the model respectively), they won't show up in the GUI if they don't have the correct contact types (they are filtered out by the proxy filter).