Provide a method for grouping the contacts together
This takes QList<uri> of nco:Contacts, creates a pimo:Person
and adds the contacts as groundingOccurences.

Removing person is done by removing pimo:Person from Nepomuk,
which will be signalled by our resource watcher, which will
remove it from the model as well. The contacts stay intact
in both Nepomuk and model.