Clean the query results processing and add comments
This could still use optimizations though, loading ~1000 contacts
takes now roughly 3 seconds, which is not bad, but still too much.
Might be worthy doing this in a non-blocking thread and adding
the contacts to the model continuously. That would result in immediate
GUI visibility and also visible indication of data loading.