First draft of proxy model adding Telepathy presences
The idea is to provide presence and capabilities information directly
from Telepathy, which gives us instant changes and one less layer
for the data to go through and even easing the Nepomuk load. This
is done in a proxy model, so if we don't need the real-time presence
updates, we can use the PersonsModel directly.