PersonsModel: Remove the explicit check for nepomuk:/me
Add the check in the query instead. It should be faster this way. Though
the <nepomuk:/me> should never actually be a nco:PersonContact, so this
case should never happen.

We cannot do trick for the pimo groundingOccurrence as that is in an
optinal block and it results in the query results containing
<nepomuk:/me> as the groundingOccurrence being returned except with
?pimo_groundingOccurrence not being bound to anything!