PersonsModel: Always fetch contacts which have im accounts
The PersonsModel is *for now* only going to be used in Telepathy, where
we only care about contacts with IMAccounts.

We could also just filter out the contacts which have an imAccount but
that results in us loading more contacts than required, and is a waste
of resources. The difference can be about 1000 vs 7000 contacts.

Reviewed By: Martin Klapetek