Simplify semantics in internal contact data feeding API
Change the ContactItem internals to store a QMap<int, QVariant> instead of
a multi hash. This way the type of the role's data will be always the same
 that has been fed and this will be controlled by whoever calls
setContactData (so it requires documenting the roles). Also make sure that
when we change some of the data, it still has the same type as the former.

Rename ContactItem::addData into ContactItem::setContactData. addData
could have meant that the provided data could extend the value of the
set value. I think it's better explained with the current data.

REVIEW: 109598