Use indexes directly in duplicates finder + add model method for finding duplicates
The duplicates finder right now stores only the row of the matched
contact. Whenever we want use that, we need to create an index from it.
This stores the indexes directly and emits the matched indexes. The app
should listen to this new signal and should display it to the user for
confirmation. Once user confirms, it will take the indexes that user
confirmed and pass them to createPersonFromIndexes(..) method, this will
be part 2 of this patch.

REVIEW: 110134