Add only contacts from the watcher that we are interested in
If we request IM only contacts from PersonsModel and then akonadi feeder
kicks in and starts feeding its contacts, resource watcher catches that
and puts new contacts in the model. This wasn't a problem before as we
used a proxy model to filter out contacts we don't want. With the
Features in place this is a problem again.

This patch removes the contactCreated slot, which put the contact in the
model, and instead puts the contact in the model while we receive
propertyChanged() signal. This is because the resource watcher "works"
in two phases - each new created contact is signaled first and then for
each new created contacts its properties are signaled as
propertyChanged(). With propertyChanged() we can check what type of
property is being added (this information is not present in
contactCreated()) and if that property is the one we are interested in
(defined by the features) only then we create the new contact in the
model.

REVIEW: 110128