Removing PersonsPresenceModel; switching to IMPresencePlugin
This disables the PersonsPresenceModel and uses IMPresencePlugin for
presence data instead. Also the translation proxy now implements KTp
stuff directly, will be moved to KTp eventually (same for the plugins).

This also adds a temp workaround for linking against the
IMPresencePlugin - the cpp file is added to kpeople sources. This will
be solved when we move it to KTp.

PersonItem now always returns a list for all data except of
Qt::DisplayRole and Qt::DecorationRole.

Currently we miss a contact update system - the presence plugin emits
contactUpdated() with Tp id, we need to translate this to some model
index and emit dataUpdated() for that index.

Reviewed-by: David Edmundson