Group multiple contact fields on the server
Normally a contact with multiple email addresses (for example)
will appear as multiple rows in the query.

Currently our model does not handle this and multiple emails/phones
etc. are missing in the model

By using group_digest we group all values into a single list on the
server

REVIEW: 110817