Reenable contact merging
For contact merging we need the contact uris and for that we need a way
to retrieve it from the fake persons (which represents the contact).
Since the model will be used with proxies, we can't guarantee that
index.child(0,0).data(...) will work, so we need another way to access
that contact's uri. So I added new role called ChildContactUriRole,
which returns the child contact role and it's meant to be used only with
fake persons.

I'm not sure this is a good idea, needs proper reviewing.