Add a nepomuk query to get Tp::Account from simple contact id
If the KTp contact is null, we still need the Tp account for that
contact so we can either group it properly or bring that account online
if user starts a chat with a contact that belongs to offline account.

Nepomuk stores the account identifier, so we can retrieve it from
Nepomuk directly (it's impossible to retrieve it from Telepathy if that
account is offline).