Merging contacts dialog and the example to make it run
This dialog shows all your contacts that have duplicates.
The list view allows to see the name of the contact who is
concerned by the merge and the photos of the related matching
contacts. When click on the item of the list, the view extends
to show merging contacts details : name+merging reason(ex:mail)
The effective merge is not currently working due to a known issue :
Nepomuk service storage issue, so it's commented untill it get fixed

Review Request #111165