Make contacts merging/unmerging methods static
The reason being that not always the client application has access to
the instantiated PersonsModel as it might be hidden behind proxies and
only the proxy can be passed to the view/client app. For these cases,
it's very convenient to have these methods as static members, so we can
call it from wherever it's needed.

I've also made the createPersonFromUris return a KJob, which can be
tracked on the client side, it's also needed to obtain the new person's
uri and set the view to that uri.

REVIEW: 111386