Improve Merge Dialog interface
Ports the dialog to use a KPixmapSequenceWidget that will be displayed when
items are being merged. The dialog doesn't close until it's done.
To do so, we use the MatchesSolver instead of calling the matching from
within the dialog and disregarding the evolution.
Less code, more features.

CCMAIL: franck.arrecot@gmail.com