Make All Data Loading Async
Plugins now return two objects: AllContactsMonitor and/or ContactMonitor
which monitor all contact changes, or load only a singe contact
respectively

These are ref counted to save resources.

A default ContactMonitor is provided that uses the AllContactsMonitor
filtering only the relevant information.

The Akonadi plugin is ported to this new system.