Make PersonsModel a tree model
Turns out converting a list model into a tree model on a proxy level is
not as easy without fully reimplementing most of the proxy model.

So we did it the other way around - make PersonsModel a tree model
(internally it already is anyway) which can be flatted later if needed.