Fix 'attempting to add QLayout to widget which already has a layout' warning
Setting parent widget to a layout automatically tries to set the new
layout as widget's main layout. In this case however we want to insert
the layout into the existing main layout (created above), so should not
set the parent widget. addLayout() takes ownership of the inserted layout,
so the QFormLayout won't be leaked.