Insert contacts only inside beginInsertRows/endInsertRows calls
The original idea of inserting the contact first, getting the index and
then just emitting the proper signals causes the view/proxy model to add
one ghost contact entry because the beginInsertRows call makes the
view/proxy model to reserve one new entry, but that empty spot is never
filled as the contact has already been inserted before the call.

So now we always just append it at the end. Fixes the ghost contact
problem.

Reviewed-by: David Edmundson