Search AllContactsMonitor first for a contact when requesting ContactMonitor
If we have the whole model already loaded, we have all the data we want,
so all we have to do is query the model (AllContactsMonitor) and if it
does have the contact we're looking for, just retrieve it and create
ContactMonitor with it. This will bypass issuing the query to the
backend to obtain the data (which we already have).