Reverse the filtering logic in PersonsSortFilterProxyModel
For filtering multiple properties the logic needs to be OR
rather than AND.

For example in Plasma Mobile, one can start a conversation
either with a phonenumber contact or IM contact, so both
properties need to be requested. But with AND matching it
won't return anything (or only those contacts that have both
and user merged them).

REVIEW: 127117