Make PersonPluginManager API public
Summary: This patch makes PersonPluginManager available to third-party applications, allowing them to use KPeople in a stand-alone way, without interacting with system-wide plugins.

It is part of {T9595}. Currently we only want to use only Akonadi plugin (to be developed), and then possibly add support for other (i.e. KTelepathy) plugins.
For Akonadi plugin, we want to pass an EntityTreeModel to use it as data source, which would be impossible if we use system-wide plugin system.

Test Plan: It compiles, it works.

Reviewers: #frameworks, dvratil, apol

Reviewed By: dvratil

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D15431