Fix build with Qt 5.15
In 88f8e64618cf addresseeMap was a KContacts::Addressee::Map and was changed
to a QMap, the KContacts::Addressee::Map documentation says that the keys
are unique.

And it looks like the keys in a map returned by contacts() are unique in each
map, and each plugin uses keys that are urls specific to it (so the akonadi
plugin uses akonadi urls and the ktp plugin uses ktp urls); thanks to dfaure
for figuring that out.

NO_CHANGELOG