Add avatar image provider
We need a way to integrate KPeople with Kirigami.Avatar. Kirigami.Avatar
uses a QtQuick.Image internally, which can not directly have a QPixmap
assigned to its source property.
Changing Kirigami.Avatar to use Kirigami.Icon is not viable, since it
already exposes some Image internals.

The official way to display a QPixmap in an image is using image
providers, but we are in an unfortunate situation, because we are
already using uris, and image providers do too.
The new image provider understands base64 encoded person uris, and
convinience methods to create these from QML are provided.