vcard: remove plugin id
The KPlugin documentation advises that C++ plugins should not set the
id field explicitly, as it will be set from the filename, and since
be10ddbf an explicit error message is shown indicating that this id
should be removed. The json file for the vcard backend contains an id
field though, leading to an unnecessary warning every time it is loaded
(e.g. on kdeconnect-sms startup).

This change removes the unnecessary id field.